+++ src/popup/popup.js
@@ -1,3 +1,5 @@
+let enableRegex;
+
 function main() {
 
     const $loginbutton = document.getElementById('login-button');
@@ -7,9 +9,17 @@ function main() {
     const $logoutbutton = document.getElementById('logout-button');
     $logoutbutton.addEventListener('click', handleOnClickLogout);
 
+    const $updatebutton = document.getElementById("update-filter-button");
+    $updatebutton.addEventListener('click', handleOnClickUpdateFilter);
+
     document.getElementById("email-input-form").addEventListener("keypress", submitLoginOnKeyEnter);
     document.getElementById("password-input-form").addEventListener("keypress", submitLoginOnKeyEnter);
     document.getElementById("idp-input-form").addEventListener("keypress", submitLoginOnKeyEnter);
+    document.getElementById("domain-input-form").addEventListener("keypress", submitLoginOnKeyEnter);
+
+    document.getElementById("enable-regex-button").addEventListener('click', handleOnClickRegex)
+
+    enableRegex = false;
 
     chrome.runtime.sendMessage({
         msg: "check-authenticated"
@@ -47,12 +57,13 @@ async function handleOnClickLogin(event) {
     let email = document.getElementById("email-input-form").value
     let password = document.getElementById("password-input-form").value
     let idp = document.getElementById("idp-input-form").value
+    let filter = document.getElementById("domain-input-form").value
 
     if (!(email && password && idp)) {
         return;
     }
 
-    document.getElementById("loader").classList.remove('hidden');
+    document.getElementById("login-loader").classList.remove('hidden');
     document.getElementById("generate-button-text").classList.add('hidden');
     document.getElementById("login-status-fail").classList.add('hidden');
 
@@ -65,6 +76,8 @@ async function handleOnClickLogin(event) {
         email,
         password,
         idp,
+        filter,
+        regex: enableRegex
     }, function (response) {
         handleAfterLoginAttempt(response)
     });
@@ -80,12 +93,37 @@ function handleOnClickLogout() {
 
     document.getElementById("login-button").classList.remove("d-none");
     document.getElementById('logout-button-div').classList.add("d-none");
-    document.getElementById("credential-input-forms").classList.remove('hidden');
+    document.getElementById("update-filter-button").classList.add("d-none");
     document.getElementById("login-status-success").classList.add('hidden');
+
+    document.getElementById("email-input-div").classList.remove('hidden');
+    document.getElementById("password-input-div").classList.remove('hidden');
+    document.getElementById("idp-input-div").classList.remove('hidden');
 
     document.getElementById("email-input-form").value = '';
     document.getElementById('password-input-form').value = '';
     document.getElementById('idp-input-form').value = '';
+    document.getElementById('domain-input-form').value = '';
+    document.getElementById("enable-regex-button").classList.remove("toggled");
+
+    enableRegex = false;
+}
+
+/**
+ * Handle user's request to update domain filter
+ */
+function handleOnClickUpdateFilter() {
+    document.getElementById("update-loader").classList.remove('hidden');
+    document.getElementById("update-button-text").classList.add('hidden');
+    let filter = document.getElementById("domain-input-form").value
+    chrome.runtime.sendMessage({
+        msg: "update-filter",
+        filter,
+        regex: enableRegex
+    }, function () {
+        document.getElementById("update-loader").classList.add('hidden');
+        document.getElementById("update-button-text").classList.remove('hidden');
+    });
 }
 
 /**
@@ -95,20 +133,52 @@ function handleOnClickLogout() {
 function handleAfterLoginAttempt(options) {
     const {success, error} = options;
 
-    document.getElementById("loader").classList.add('hidden');
+    document.getElementById("login-loader").classList.add('hidden');
     document.getElementById("generate-button-text").classList.remove('hidden');
 
     if (success) {
-        document.getElementById("credential-input-forms").classList.add('hidden');
+        document.getElementById("email-input-div").classList.add('hidden');
+        document.getElementById("password-input-div").classList.add('hidden');
+        document.getElementById("idp-input-div").classList.add('hidden');
+
         document.getElementById("login-status-success").classList.remove('hidden');
         document.getElementById("login-status-fail").classList.add('hidden');
         document.getElementById("login-button").classList.add("d-none");
         document.getElementById('logout-button-div').classList.remove("d-none");
+        document.getElementById("update-filter-button").classList.remove("d-none");
+
+        chrome.runtime.sendMessage({
+            msg: "get-filter"
+        }, function (response) {
+            document.getElementById("domain-input-form").value = response.domainFilter;
+            const $regexButton = document.getElementById("enable-regex-button")
+
+            enableRegex = response.enableRegex
+
+            if (response.enableRegex) {
+                $regexButton.classList.add("toggled");
+            } else {
+                $regexButton.classList.remove("toggled");
+            }
+        });
     } else {
         document.getElementById('login-fail-error').innerText = error + '.';
         document.getElementById("login-status-fail").classList.remove('hidden');
         document.getElementById("login-status-success").classList.add('hidden');
     }
+}
+
+/**
+ * Handle user's request to enable or disable regex syntax on domain filter
+ */
+function handleOnClickRegex() {
+    const $regexButton = document.getElementById("enable-regex-button")
+    if (enableRegex) {
+        $regexButton.classList.remove("toggled");
+    } else {
+        $regexButton.classList.add("toggled");
+    }
+    enableRegex = !enableRegex;
 }
 
 main();